+++ src/iptux/CoreThreadTest.cpp
@@ -1,8 +1,12 @@
 #include "gtest/gtest.h"
 
+#include <thread>
+
 #include "iptux/CoreThread.h"
 #include "iptux/TestHelper.h"
 #include "iptux/utils.h"
+#include "iptux/Exception.h"
+#include "iptux/output.h"
 
 using namespace std;
 using namespace iptux;
@@ -37,10 +41,10 @@ TEST(CoreThread, GetPalList) {
   auto core = make_shared<ProgramData>(config);
   core->sign = "abc";
   CoreThread* thread = new CoreThread(core);
-  EXPECT_EQ(thread->GetPalList(), nullptr);
-  PalInfo pal;
-  thread->AttachPalToList(&pal);
-  EXPECT_NE(thread->GetPalList(), nullptr);
+  EXPECT_EQ(int(thread->GetPalList().size()), 0);
+  PPalInfo pal = make_shared<PalInfo>();
+  thread->AttachPalToList(pal);
+  EXPECT_EQ(int(thread->GetPalList().size()), 1);
   delete thread;
 }
 
@@ -49,7 +53,15 @@ TEST(CoreThread, SendMessage) {
   auto core = make_shared<ProgramData>(config);
   core->sign = "abc";
   CoreThread* thread = new CoreThread(core);
-  PalInfo pal;
+  auto pal = make_shared<PalInfo>();
+  try {
+    thread->SendMessage(pal, "hello world");
+    EXPECT_TRUE(false);
+  } catch(Exception& e) {
+    EXPECT_EQ(e.getErrorCode(), ErrorCode::PAL_KEY_NOT_EXIST);
+  }
+
+  thread->AttachPalToList(pal);
   EXPECT_TRUE(thread->SendMessage(pal, "hello world"));
   delete thread;
 }
@@ -59,7 +71,8 @@ TEST(CoreThread, SendMessage_ChipData) {
   auto core = make_shared<ProgramData>(config);
   core->sign = "abc";
   CoreThread* thread = new CoreThread(core);
-  PalInfo pal;
+  auto pal = make_shared<PalInfo>();
+  thread->AttachPalToList(pal);
   ChipData chipData;
   chipData.data = "hello world";
   EXPECT_TRUE(thread->SendMessage(pal, chipData));
@@ -71,11 +84,12 @@ TEST(CoreThread, SendMsgPara) {
   auto core = make_shared<ProgramData>(config);
   core->sign = "abc";
   CoreThread* thread = new CoreThread(core);
-  PalInfo pal;
+  PPalInfo pal = make_shared<PalInfo>();
+  thread->AttachPalToList(pal);
   ChipData chipData;
   chipData.data = "hello world";
   MsgPara para;
-  para.pal = &pal;
+  para.pal = pal;
   para.dtlist.push_back(move(chipData));
   EXPECT_TRUE(thread->SendMsgPara(para));
   delete thread;
@@ -102,7 +116,53 @@ TEST(CoreThread, SendAskShared) {
   auto core = make_shared<ProgramData>(config);
   core->sign = "abc";
   CoreThread* thread = new CoreThread(core);
-  PalInfo pal;
+  auto pal = make_shared<PalInfo>();
   thread->SendAskShared(pal);
   delete thread;
+}
+
+TEST(CoreThread, FullCase) {
+  using namespace std::chrono_literals;
+  auto oldLogLevel = Log::getLogLevel();
+  Log::setLogLevel(LogLevel::DEBUG);
+  auto config1 = IptuxConfig::newFromString("{\"bind_ip\": \"127.0.0.1\"}");
+  auto thread1 = new CoreThread(make_shared<ProgramData>(config1));
+  thread1->start();
+  auto config2 = IptuxConfig::newFromString("{\"bind_ip\": \"127.0.0.2\"}");
+  auto thread2 = new CoreThread(make_shared<ProgramData>(config2));
+  thread2->start();
+  thread1->SendDetectPacket("127.0.0.2");
+  while(thread2->GetOnlineCount() != 1) {
+    this_thread::sleep_for(10ms);
+  }
+
+  EXPECT_TRUE(thread2->GetPal("127.0.0.1"));
+  EXPECT_EQ(thread1->GetOnlineCount(), 1);
+  EXPECT_TRUE(thread1->GetPal("127.0.0.2"));
+
+  vector<shared_ptr<const Event>> thread2Events;
+  thread2->registerCallback([&](shared_ptr<const Event> event) { thread2Events.emplace_back(event); });
+
+  auto pal2InThread1 = thread1->GetPal("127.0.0.2");
+  auto pal1InThread2 = thread2->GetPal("127.0.0.1");
+  thread1->SendMessage(pal2InThread1, "hello world");
+  while(thread2Events.size() != 1) {
+    this_thread::sleep_for(10ms);
+  }
+  auto event = thread2Events[0];
+  EXPECT_EQ(event->getType(), EventType::NEW_MESSAGE);
+  auto event2 = (NewMessageEvent*)(event.get());
+  EXPECT_EQ(event2->getMsgPara().dtlist[0].ToString(), "ChipData(MessageContentType::STRING, hello world)");
+
+  thread1->SendExit(pal2InThread1);
+  while(thread2->GetOnlineCount() != 0) {
+    this_thread::sleep_for(10ms);
+  }
+
+  thread1->SendDetectPacket("127.0.0.2");
+  while(thread2->GetOnlineCount() != 1) {
+    this_thread::sleep_for(10ms);
+  }
+
+  Log::setLogLevel(oldLogLevel);
 }