+++ src/iptux/Event.h
@@ -7,6 +7,7 @@ namespace iptux {
 enum class EventType {
   NEW_PAL_ONLINE,
   NEW_MESSAGE,
+  PAL_OFFLINE,
 };
 
 class Event {
@@ -15,27 +16,32 @@ class Event {
   virtual ~Event() = default;
 
   EventType getType() const;
-  virtual Event* clone() const = 0;
  private:
   EventType type;
 };
 
 class NewPalOnlineEvent:public Event {
  public:
-  explicit NewPalOnlineEvent(PalInfo* palInfo);
-  const PalInfo* getPalInfo() const;
-  NewPalOnlineEvent* clone() const override;
+  explicit NewPalOnlineEvent(PPalInfo palInfo);
+  CPPalInfo getPalInfo() const;
  private:
-  PalInfo* palInfo;
+  PPalInfo palInfo;
 };
 
 class NewMessageEvent:public Event {
  public:
   explicit NewMessageEvent(MsgPara&& msgPara);
   const MsgPara& getMsgPara() const;
-  NewMessageEvent* clone() const override;
  private:
   MsgPara msgPara;
+};
+
+class PalOfflineEvent: public Event {
+ public:
+  explicit PalOfflineEvent(PalKey palKey);
+  const PalKey& GetPalKey() const;
+ private:
+  PalKey palKey;
 };
 
 }