+++ src/iptux/UdpData.h
@@ -38,7 +38,7 @@ class UdpData {
   void SomeoneLost();
   void SomeoneEntry();
   void SomeoneExit();
-  void SomeoneAnsentry();
+  void SomeoneAnsEntry();
   void SomeoneAbsence();
   void SomeoneSendmsg();
   void SomeoneRecvmsg();
@@ -47,7 +47,7 @@ class UdpData {
   void SomeoneSendSign();
   void SomeoneBcstmsg();
 
-  PalInfo *CreatePalInfo();
+  std::shared_ptr<PalInfo> CreatePalInfo();
   void UpdatePalInfo(PalInfo *pal);
 
   void InsertMessage(PalInfo *pal, GroupBelongType btype, const char *msg);