+++ src/iptux/UiCoreThread.h
@@ -51,14 +51,12 @@ class UiCoreThread: public CoreThread {
   void InsertMessage(MsgPara&& para);
 
   void InsertMsgToGroupInfoItem(GroupInfo *grpinf, MsgPara *para);
-  static void SendFeatureData(PalInfo *pal);
-  static void SendBroadcastExit(PalInfo *pal, UiCoreThread* self);
-  static void UpdateMyInfo();
 
   void ClearAllPalFromList() override ;
   void DelPalFromList(PalKey palKey) override ;
   void UpdatePalToList(PalKey palKey) override ;
   void AttachPalToList(PalInfo *pal) override ;
+  void AttachPalToList(std::shared_ptr<PalInfo> pal) override ;
   GroupInfo *GetPalRegularItem(PalInfo *pal);
   GroupInfo *GetPalSegmentItem(PalInfo *pal);
   GroupInfo *GetPalGroupItem(PalInfo *pal);